More notes
The script derivative_figure_generation.jl doesn't work and not sure why, seemingly it would be something to do with the checkpointed time integration function as this should be the only difference between it and run_checkpointing.jl, which is *still* working, as of this point. A derivative check was done in run_checkpointing.jl and the derivatives look correct. Moving on to running a true model....